move init.php to interfaces dir modify composer require extension
