+++ composer.json
@@ -11,7 +11,10 @@
     "minimum-stability": "stable",
     "require": {
         "php": ">=7.1",
-        "ext-json": ">=1.2.1"
+        "ext-swoole": ">=4.0"
+    },
+    "require-dev": {
+        "eaglewu/swoole-ide-helper": "dev-master"
     },
     "autoload":{
         "psr-4":{
