+++ src/Interfaces/Init.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace core;
+namespace core\interfaces;
 
 interface Init
 {
